fix few bugs
1. can't save config json file if not exist before
2. redirect ftp server can't be set if no config before
3. UDP bin folder shows wrong one

+++ web/settings.py
@@ -180,7 +180,8 @@ class ConfigLoader(object) :
             
             # rename file
             try:
-                os.remove(clz.CONFIG_FILE)
+                if os.path.exists(clz.CONFIG_FILE):
+                    os.remove(clz.CONFIG_FILE)
                 os.rename(tmp_file, clz.CONFIG_FILE)
             except:
                 pass
